typecast to int when the variable is int!

+++ lib/progress.c
@@ -366,7 +366,7 @@ int Curl_pgrsUpdate(struct connectdata *conn)
      (data->progress.ulspeed>0) &&
      (data->progress.size_ul > 100) ) {
     ulestimate = (long)(data->progress.size_ul / data->progress.ulspeed);
-    ulpercen = (long)(100*(data->progress.uploaded/100) /
+    ulpercen = (int)(100*(data->progress.uploaded/100) /
                       (data->progress.size_ul/100) );
   }
 
@@ -375,7 +375,7 @@ int Curl_pgrsUpdate(struct connectdata *conn)
      (data->progress.dlspeed>0) &&
      (data->progress.size_dl>100)) {
     dlestimate = (long)(data->progress.size_dl / data->progress.dlspeed);
-    dlpercen = (long)(100*(data->progress.downloaded/100) /
+    dlpercen = (int)(100*(data->progress.downloaded/100) /
                       (data->progress.size_dl/100));
   }
